feat: sanitize user text entry
